JUMP

+++ src/ch/hevs/gdx2d/Frog.scala
@@ -9,7 +9,7 @@ import com.badlogic.gdx.math.Vector2
 class Frog(var posit : Vector2) extends AnimatedObject(posit) {
   override var img: BitmapImage = new BitmapImage("data/images/frog_arrow.png")
   var r : Int = 30
-  var direction : Int = 0 //degree
+  var direction : Float = 0 //degree
   var onLily : Boolean = true //rotation on
 
   def onGraphicsRender(g: GdxGraphics): Unit = {
